requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
Pillow==10.1.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
timm==1.0.20
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
